I have removed the need for the Constants file
I have removed the need for the Constants file by making the initial
settings configuration easily managed through the static "initialise"
method in the appDelegate.  I've also put the AdMob banner on a backing
view so that it can easily remain centred in both orientations.  Also
this backing layer allows you to specify a background colour for the
layer because the banner does not stretch across the full screen in
landscape mode.